New challenges

+++ src/main/java/coding/minmaxsum/MinMaxSumI.java
@@ -1,4 +1,7 @@
 package coding.minmaxsum;
+
+import java.math.BigInteger;
+
 /**
  * Given five positive integers, find the minimum and maximum values that can be calculated by summing exactly four of the five integers. 
  * Then print the respective minimum and maximum values as a single line of two space-separated long integers.For example, arr=[1,3,5,7,9]. 
@@ -7,5 +10,5 @@ package coding.minmaxsum;
  *
  */
 public interface MinMaxSumI {
-	long[] miniMaxSum(int[] arr);
+	BigInteger[] miniMaxSum(int[] arr);
 }
